+++ src/compute/MTCompute/computeMt-stdthread.cpp
@@ -98,7 +98,7 @@ vector2D LSH::computeCollision_stdthread(vector2D hMatrixN, vector2D hMatrixQ){
                             int n = i%inloop;
                             for (int hash_id = 0; hash_id < L; ++hash_id) {
                                 //cast to int for comparision, reduce double compare risk
-                                if ((int)hMatrixN[n][hash_id] == (int)hMatrixQ[q][hash_id])
+                                if (hMatrixN[n][hash_id] == hMatrixQ[q][hash_id])
                                     collisionMatrix[q][n]++;
                             }
