+++ src/compute/compute.cpp
@@ -4,6 +4,7 @@
 
 #include "../../include/LSH.h"
 #include <random>
+#include <iostream>
 
 vector2D LSH::computeHash(vector2D dataset, size_t pNum){
     vector2D hashMatrix;
@@ -43,7 +44,7 @@ vector2D LSH::computeCollision(vector2D hMatrixN, vector2D hMatrixQ){
         for (int n = 0; n <N ; ++n){
             for (int hash_id = 0; hash_id < L; ++hash_id) {
                 //cast to int for comparision, reduce double compare risk
-                if ((int)hMatrixN[n][hash_id] == (int)hMatrixQ[q][hash_id])
+                if (hMatrixN[n][hash_id] == hMatrixQ[q][hash_id])
                     singleLine[n]++;
             }
         }
